+++ src/Fund.t.sol
@@ -1,11 +1,10 @@
-pragma solidity ^0.4.17;
+pragma solidity ^0.4.19;
 
 import "ds-test/test.sol";
 import "./pricefeeds/PriceFeed.sol";
 import "./assets/PreminedAsset.sol"; import "./assets/EtherToken.sol";
 import "./exchange/thirdparty/SimpleMarket.sol";
-import "./sphere/Sphere.sol";
-import "./participation/Participation.sol";
+import "./compliance/Compliance.sol";
 import "./riskmgmt/RiskMgmt.sol";
 import "./Fund.sol";
 
@@ -14,11 +13,10 @@ contract FundTest is DSTest {
     PriceFeed datafeed;
     EtherToken etherToken;
     Fund fund;
-    Participation participation;
+    Compliance participation;
     PreminedAsset melonToken;
     RiskMgmt riskManagement;
     SimpleMarket simpleMarket;
-    Sphere sphere;
 
     // constants
     string FUND_NAME = "My Fund";
@@ -47,8 +45,7 @@ contract FundTest is DSTest {
     //    etherToken = new EtherToken();
     //    datafeed = new PriceFeed(melonToken, INTERVAL, VALIDITY);
     //    simpleMarket = new SimpleMarket();
-    //    sphere = new Sphere(datafeed, simpleMarket);
-    //    participation = new Participation();
+    //    participation = new Compliance();
     //    fund = new Fund(
     //        MANAGER_ADDRESS,
     //        FUND_NAME,
@@ -58,19 +55,20 @@ contract FundTest is DSTest {
     //        melonToken,
     //        participation,
     //        riskManagement,
-    //        sphere
+    //        datafeed,
+    //        simpleMarket
     //    );
     //}
 
     //function testVariablesSetCorrectly() {
     //    var returnedName = fund.getName();
     //    uint returnedDecimals = fund.getDecimals();
-    //    var (returnedDatafeed, returnedExchange, returnedParticipation, returnedRiskMgmt) = fund.getModules();
+    //    var (returnedDatafeed, returnedExchange, returnedCompliance, returnedRiskMgmt) = fund.getModules();
     //    uint stake = fund.getStake();
 
     //    assertEq(returnedDatafeed, datafeed);
     //    assertEq(returnedExchange, simpleMarket);
-    //    assertEq(returnedParticipation, participation);
+    //    assertEq(returnedCompliance, participation);
     //    assertEq(returnedRiskMgmt, riskManagement);
     //    //assertEq(returnedName, FUND_NAME); //TODO: uncomment when assertEq implemented for strings
     //    assertEq(returnedDecimals, MELON_DECIMALS);