+++ src/dependencies/DBC.sol
@@ -1,4 +1,4 @@
-pragma solidity ^0.4.17;
+pragma solidity ^0.4.19;
 
 /// @title Desing by contract (Hoare logic)
 /// @author Melonport AG <team@melonport.com>