+++ src/libraries/safeMath.sol
@@ -1,4 +1,4 @@
-pragma solidity ^0.4.17;
+pragma solidity ^0.4.19;
 
 /// @title Math operations with safety checks
 /// @author Melonport AG <team@melonport.com>