+++ src/system/Governance.sol
@@ -1,4 +1,4 @@
-pragma solidity ^0.4.17;
+pragma solidity ^0.4.19;
 
 import 'ds-group/group.sol';
 import '../dependencies/DBC.sol';